+++ eoreader/products/optical/s3_product.py
@@ -463,6 +463,23 @@ class S3Product(OpticalProduct):
         """
         raise NotImplementedError("This method should be implemented by a child class")
 
+    def _manage_nodata(self, band_arr: XDS_TYPE, band: obn, **kwargs) -> XDS_TYPE:
+        """
+        Manage only nodata pixels
+
+        Args:
+            band_arr (XDS_TYPE): Band array
+            band (obn): Band name as an OpticalBandNames
+            kwargs: Other arguments used to load bands
+
+        Returns:
+            XDS_TYPE: Cleaned band array
+        """
+        # Get nodata mask
+        no_data = np.where(np.isnan(band_arr.data), self._mask_true, self._mask_false)
+
+        return self._set_nodata_mask(band_arr, no_data)
+
     def _load_bands(
         self,
         bands: list,