+++ eoreader/products/optical/s3_slstr_product.py
@@ -43,8 +43,9 @@ from eoreader.bands import ALL_CLOUDS, CIRRUS, CLOUDS, RAW_CLOUDS, BandNames
 from eoreader.bands import OpticalBandNames as obn
 from eoreader.bands import to_str
 from eoreader.exceptions import InvalidProductError, InvalidTypeError
-from eoreader.keywords import SLSTR_RAD_ADJUST, SLSTR_STRIPE, SLSTR_VIEW
+from eoreader.keywords import CLEAN_OPTICAL, SLSTR_RAD_ADJUST, SLSTR_STRIPE, SLSTR_VIEW
 from eoreader.products import S3DataType, S3Instrument, S3Product, S3ProductType
+from eoreader.products.optical.optical_product import CleanMethod
 from eoreader.reader import Platform
 from eoreader.utils import EOREADER_NAME
 
@@ -763,8 +764,6 @@ class S3SlstrProduct(S3Product):
 
         return e0
 
-    # pylint: disable=R0913
-    # R0913: Too many arguments (6/5) (too-many-arguments)
     def _manage_invalid_pixels(
         self, band_arr: XDS_TYPE, band: obn, **kwargs
     ) -> XDS_TYPE:
@@ -1001,9 +1000,13 @@ class S3SlstrProduct(S3Product):
         Returns:
             Union[CloudPath, Path]: Clean band path
         """
+        cleaning_method = CleanMethod.from_value(
+            kwargs.get(CLEAN_OPTICAL, CleanMethod.CLEAN)
+        )
+
         suffix = self._get_suffix(band, **kwargs)
         res_str = self._resolution_to_str(resolution)
 
         return self._get_band_folder(writable).joinpath(
-            f"{self.condensed_name}_{band.name}_{suffix}_{res_str.replace('.', '-')}_clean.tif",
+            f"{self.condensed_name}_{band.name}_{suffix}_{res_str.replace('.', '-')}_{cleaning_method.value}.tif",
         )